Fix on psector_vdefault on fct_audit_check_project

+++ dbmodel/utils/fct/gw_fct_audit_check_project.sql
@@ -29,6 +29,7 @@ query_string text;
 max_aux int8;
 project_type_aux text;
 rolec_rec record;
+psector_vdef_aux text;
 
 BEGIN 
 
@@ -41,8 +42,10 @@ BEGIN
 	enabled_bool:=FALSE;
 	count=0;
 
+	SELECT value INTO psector_vdef_aux FROM config_param_user WHERE parameter='psector_vdefault' AND cur_user=current_user;
+	
 	-- Force psector vdefault visible to current_user (only to => role_master)
-	IF 'role_master' IN (SELECT rolname FROM pg_roles WHERE  pg_has_role( current_user, oid, 'member')) THEN
+	IF 'role_master' IN (SELECT rolname FROM pg_roles WHERE  pg_has_role( current_user, oid, 'member')) and  psector_vdef_aux is not null THEN
 	  DELETE FROM selector_psector WHERE psector_id =(SELECT value FROM config_param_user WHERE parameter='psector_vdefault' AND cur_user=current_user)::integer AND cur_user=current_user;
 	  INSERT INTO selector_psector (psector_id, cur_user) VALUES ((SELECT value FROM config_param_user WHERE parameter='psector_vdefault' AND cur_user=current_user)::integer, current_user);
 	END IF;
